requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
asyncpg==0.29.0
pydantic==2.7.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
httpx==0.27.0
python-dotenv==1.0.0
stripe==9.12.0
